Fin application + logo

+++ app/src/main/java/com/lille1/hvalette/gestionparclille1/activity/DetailActivity.kt
@@ -1,5 +1,7 @@
-package com.lille1.hvalette.gestionparclille1
+package com.lille1.hvalette.gestionparclille1.activity
 
+import android.app.AlertDialog
+import android.content.DialogInterface
 import android.location.Geocoder
 import android.net.Uri
 import android.support.v7.app.AppCompatActivity
@@ -8,7 +10,9 @@ import android.widget.Toast
 import kotlinx.android.synthetic.main.activity_detail.*
 import java.util.*
 import android.content.Intent
-
+import com.lille1.hvalette.gestionparclille1.GestionParcLille1App
+import com.lille1.hvalette.gestionparclille1.R
+import com.lille1.hvalette.gestionparclille1.entity.Probleme
 
 
 class DetailActivity : AppCompatActivity() {
@@ -16,9 +20,10 @@ class DetailActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_detail)
+
         val id = this.intent.extras.get("id")
         val probleme = GestionParcLille1App.database.problemeDao().getById(id.toString().toInt())
-        Toast.makeText(this, "" + id, Toast.LENGTH_SHORT)
+
         when (probleme.type) {
             0 -> {
                 typeImage.setImageResource(R.drawable.tree)
@@ -41,22 +46,48 @@ class DetailActivity : AppCompatActivity() {
                 type.text = "Mauvaises herbes à enlever"
             }
         }
+
         description.text = probleme.description
-        val geocoder = Geocoder(this, Locale.getDefault())
-        val addresses = geocoder.getFromLocation(probleme.latitude, probleme.longitude, 1)
-        addresse.text = addresses.get(0).getAddressLine(0)
+        addresse.text = probleme.adresse
+
         localiser.setOnClickListener {
-            val gmmIntentUri = Uri.parse("geo:" + probleme.latitude + "," + probleme.longitude + "?q=" + addresses.get(0).getAddressLine(0))
+            val gmmIntentUri = Uri.parse("geo:" + probleme.latitude + "," + probleme.longitude + "?q=" + probleme.adresse)
             val mapIntent = Intent(Intent.ACTION_VIEW, gmmIntentUri)
             mapIntent.setPackage("com.google.android.apps.maps")
             startActivity(mapIntent)
         }
 
         supprimer.setOnClickListener {
-            GestionParcLille1App.database.problemeDao().delete(probleme)
-            val intent = Intent(this, HomeActivity::class.java)
-            startActivity(intent)
+            deleteConfirmation(probleme)
+
         }
 
+    }
+
+    /**
+     * Méthode qui montre une modale avant de supprimer le probleme
+     */
+    private fun deleteConfirmation(probleme: Probleme){
+        lateinit var dialog:AlertDialog
+        val builder = AlertDialog.Builder(this)
+        builder.setTitle("Supprimer le problème")
+        builder.setMessage("Etes vous sûr de vouloir supprimer ce problème ?")
+
+
+        // On click listener for dialog buttons
+        val dialogClickListener = DialogInterface.OnClickListener{ _, which ->
+            when(which){
+                DialogInterface.BUTTON_POSITIVE -> {
+                    GestionParcLille1App.database.problemeDao().delete(probleme)
+                    val intent = Intent(this, HomeActivity::class.java)
+                    startActivity(intent)
+                }
+            }
+        }
+
+        builder.setPositiveButton("Supprimer",dialogClickListener)
+        builder.setNegativeButton("Annuler",dialogClickListener)
+        dialog = builder.create()
+        dialog.show()
     }
 }
